+++ App_window/main_window.py
@@ -1,24 +1,28 @@
 import sys
 import sqlite3
-from pprint import pprint
 
 from PyQt6.QtWidgets import QApplication, QMainWindow, QFileDialog, QPushButton
+
 from ui.main_pageUI import Ui_MainWindow
 from PyQt6 import QtCore, QtWidgets
-from math_window import MathWindow
+from study_windows import MathWindow, PhysWindow, BioWindow, ChemWindow
+from service_windows import SettingsWindow, AccountWindow
 
 
 # Наследуемся от виджета из PyQt6.QtWidgets и от класса с интерфейсом
 class Mainwindow(QMainWindow, Ui_MainWindow):
+
+    # FIXME НЕТ ЗАПРОСОВ В БД
+
     def __init__(self):
+
         super().__init__()
         # Инициализация и обновление данных
         self.setupUi(self)
         self.progress_update()
         self.send_button.clicked.connect(self.send_article)
-        self.math_button.clicked.connect(self.open_math)
+        self.math_button.clicked.connect(self.open_study_window)
         self.senders_text.setPlaceholderText('Написать здесь...')
-
 
     def progress_update(self):
 
@@ -26,6 +30,7 @@ class Mainwindow(QMainWindow, Ui_MainWindow):
         # Выполнение запроса и получение результатов по успеваемости
         self.student_progress = con.cursor().execute(""" """).fetchall()
         # Запись успеваемости в шкалы успеваемости каждого предмета
+        # # TODO: доделать запись
         # # self.math_progression.setValue()
         # # self.phys_progression.setValue()
         # # self.bio_progression.setValue()
@@ -33,11 +38,25 @@ class Mainwindow(QMainWindow, Ui_MainWindow):
         con.close()
 
     # Функции открытия других окон
+    # # TODO: доделать окна
 
-    def open_math(self):
-        self.second_form = MathWindow(self)
+    def open_study_window(self):
+        if self.sender().text() == 'Математика':
+            self.second_form = MathWindow(self)
+        if self.sender().text() == 'Физика':
+            self.second_form = PhysWindow(self)
+        if self.sender().text() == 'Химия':
+            self.second_form = ChemWindow(self)
+        if self.sender().text() == 'Биология':
+            self.second_form = BioWindow(self)
         self.second_form.show()
-        # ex.close()
+
+    def open_service_window(self):
+        if self.sender().text() == 'Перейти в свой аккаунт':
+            self.second_form = AccountWindow(self)
+        if self.sender().text() == 'Настройки':
+            self.second_form = SettingsWindow(self)
+        ex.close()
 
     # Отправка кастомных статей по разным предметам
     def send_article(self):
@@ -57,7 +76,7 @@ class Mainwindow(QMainWindow, Ui_MainWindow):
         con.close()
 
 
-if __name__ == '__main__': #
+if __name__ == '__main__':  # запуск мейн окна
     if hasattr(QtCore.Qt, 'AA_EnableHighDpiScaling'):
         QtWidgets.QApplication.setAttribute(QtCore.Qt.AA_EnableHighDpiScaling, True)
